Fold LLMWorkerInternalAbstract into LLMWorkerAbstract (breaking)
The LLM worker family now matches the other three (img_gen, extract,
search): the abstract base holds the InferenceModelSpec
(__init__(inference_model, reporting_delegate=None)) and owns the whole
job lifecycle — llm_job_before_start, spec-driven capability checks,
constraint application, and spec-based OTel naming. Previously those
lived on the intermediate LLMWorkerInternalAbstract (a remnant of the
pre-plugin-system "external plugin" era), so an out-of-tree worker
subclassing the documented LLMWorkerAbstract silently inherited a
template method with a hole in it: forgetting to call
llm_job_before_start crashed reporting on a None duration. That state
is now unrepresentable — a plugin worker implements _gen_text /
_gen_object and nothing else.

- LLMWorkerInternalAbstract deleted; all builtin LLM workers re-parented
  onto LLMWorkerAbstract directly.
- Legacy set_llm_worker_from_external_plugin removed from the inference
  manager and its protocol (breaking): it registered spec-less worker
  classes by handle — the old meaning of "external plugin". The real
  seam (a pipelex.plugins entry point registering (family, sdk)
  backends) fully replaces it; its test now drives an out-of-tree
  LLMWorkerAbstract subclass with a real spec.
- Prompt/response debug dumps moved from the worker base to the
  llm_generate funnel; the dump helpers in llm_utils are config-gated
  internally, keeping the worker ABC free of pipelex.config (the hub
  imports the ABC at module level, and config imports the hub — a
  config import from the ABC would be an import cycle).
