+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Changed
+
+- **`LLMWorkerInternalAbstract` folded into `LLMWorkerAbstract` (breaking):** The LLM worker family now matches the other three (img_gen, extract, search): the abstract base holds the `InferenceModelSpec` (`__init__(inference_model, reporting_delegate=None)`) and owns the whole job lifecycle — `llm_job_before_start` (started_at stamp + usage-report init), spec-driven capability checks (vision, documents), constraint application, and spec-based OTel naming. Previously those lived on the intermediate `LLMWorkerInternalAbstract` (a remnant of the pre-plugin-system "external plugin" era), so an out-of-tree worker subclassing the documented `LLMWorkerAbstract` silently inherited a template method with a hole in it: forgetting to call `llm_job_before_start` crashed reporting on a `None` duration. That state is now unrepresentable — a plugin worker implements `_gen_text`/`_gen_object` and nothing else. `LLMWorkerInternalAbstract` is gone; subclasses now extend `LLMWorkerAbstract` directly. The prompt/response debug dumps moved from the worker base to the `llm_generate` funnel (the dump helpers in `llm_utils` are now config-gated internally), keeping the worker ABC free of `pipelex.config` (the hub imports the ABC at module level, and config imports the hub — a config import from the ABC would be an import cycle).
+- **Legacy `set_llm_worker_from_external_plugin` removed (breaking):** This pre-plugin-system path on the inference manager registered a spec-less worker class by handle — the old meaning of "external plugin". The real seam (a `pipelex.plugins` entry point registering `(family, sdk)` backends) fully replaces it.
+
 ### Fixed
 
 - **Routing-profile `optional_routes` were silently dropped:** `RoutingProfileFactory.make_routing_profile` parsed and validated the `optional_routes` table from `routing_profiles.toml` but never passed it to the built `RoutingProfile`, so optional routes declared in config had no effect — models relying on one fell through to the profile's default backend and could silently vanish from the model deck. The factory now carries `optional_routes` through, with unit coverage for the pass-through and for the enabled/disabled-backend gating.
