+++ pipelex/cogt/inference/inference_manager_protocol.py
@@ -18,14 +18,6 @@ class InferenceManagerProtocol(Protocol):
 
     def get_llm_worker(self, llm_handle: str) -> LLMWorkerAbstract: ...
 
-    def set_llm_worker_from_external_plugin(
-        self,
-        llm_handle: str,
-        *,
-        llm_worker_class: type[LLMWorkerAbstract],
-        should_warn_if_already_registered: bool = True,
-    ): ...
-
     ####################################################################################################
     # IMG Generation Workers
     ####################################################################################################
