+++ pipelex/cogt/llm/llm_utils.py
@@ -2,9 +2,13 @@ from typing import Any
 
 from pipelex import pretty_print
 from pipelex.cogt.llm.llm_prompt import LLMPrompt
+from pipelex.config import get_config
 
 
 def dump_prompt(llm_prompt: LLMPrompt) -> None:
+    """Print the prompt for debugging. No-op unless `cogt.llm_config.is_dump_text_prompts_enabled`."""
+    if not get_config().cogt.llm_config.is_dump_text_prompts_enabled:
+        return
     prompt_dump = ""
     if user_text := llm_prompt.user_text:
         prompt_dump += f"\n# User text:\n{user_text}\n"
@@ -20,6 +24,9 @@ def dump_prompt(llm_prompt: LLMPrompt) -> None:
 
 
 def dump_response_from_text_gen(response: Any) -> None:
+    """Print the LLM response for debugging. No-op unless `cogt.llm_config.is_dump_response_text_enabled`."""
+    if not get_config().cogt.llm_config.is_dump_response_text_enabled:
+        return
     pretty_print(response, title="Response from LLM provider")
