+++ pipelex/cogt/llm/llm_worker_factory.py
@@ -1,6 +1,6 @@
 from typing import cast
 
-from pipelex.cogt.llm.llm_worker_internal_abstract import LLMWorkerInternalAbstract
+from pipelex.cogt.llm.llm_worker_abstract import LLMWorkerAbstract
 from pipelex.cogt.model_backends.model_spec import InferenceModelSpec
 from pipelex.hub import get_inference_backend_registry, get_models_manager, get_sdk_client_manager
 from pipelex.plugins.inference_backend_registry import InferenceFamily
@@ -14,7 +14,7 @@ class LLMWorkerFactory:
         inference_model: InferenceModelSpec,
         *,
         reporting_delegate: ReportingProtocol | None = None,
-    ) -> LLMWorkerInternalAbstract:
+    ) -> LLMWorkerAbstract:
         model_handle = ModelHandle.make_for_inference_model(inference_model=inference_model)
         backend = get_models_manager().get_required_inference_backend(inference_model.backend_name)
         make_worker = get_inference_backend_registry().lookup(family=InferenceFamily.LLM, sdk=model_handle.sdk)
@@ -27,4 +27,4 @@ class LLMWorkerFactory:
         # The LLM family registry only ever holds LLM workers (the (family, sdk)
         # key guarantees it); the uniform MakeWorkerFn return type is widened to
         # InferenceWorkerAbstract, so narrow it back here.
-        return cast("LLMWorkerInternalAbstract", worker)
+        return cast("LLMWorkerAbstract", worker)
