+++ tests/integration/pipelex/cogt/test_setup_inference_workers.py
@@ -1,4 +1,3 @@
-from pipelex.cogt.llm.llm_worker_internal_abstract import LLMWorkerInternalAbstract
 from pipelex.cogt.model_backends.model_type import ModelType
 from pipelex.hub import get_inference_manager, get_model_deck
 
@@ -10,8 +9,7 @@ class TestSetupInferenceWorkers:
             match inference_model.model_type:
                 case ModelType.LLM:
                     llm_worker = inference_manager.get_llm_worker(llm_handle=model_handle)
-                    if isinstance(llm_worker, LLMWorkerInternalAbstract):
-                        assert inference_model == llm_worker.inference_model
+                    assert inference_model == llm_worker.inference_model
                 case ModelType.TEXT_EXTRACTOR:
                     _ = inference_manager.get_extract_worker(extract_handle=model_handle)
                 case ModelType.IMG_GEN:
